add fix to add chaperone bugs from the renderByRole branch merge
note: this branch should be compared to the feature-trip_detail_renderByRole branch before merging to develop
